`task group` -> `taskGroup` in code snippets
In the rename from "nursery" to "task group", a few instances of `task group` in code slipped through, which made for somewhat confusing reading (e.g. it had me wondering whether 'task' was a keyword). I've changed these all to read `taskGroup` instead. 

In one case, `group` was (validly) used as the closure parameter; I also changed that to `taskGroup` for consistency with the others and for clarity.